Polish header and hero section
Match the styling of the fonts and colors to the guide on 02.png.
Some of the colors and fonts from the previous commit did not fully
line up with the style guide.

In order to match the visual guide of the website on 01.png, more
padding and margins were applied to the elements and adjustments to
the flex containers and flex items were made in order to better match
the website to the reference image.